+++ Back end/src/main/java/com/navarro/food/navarrosfood/enums/UserRole.java
@@ -1,4 +1,4 @@
-package com.navarro.food.navarrosfood.model.Enums;
+package com.navarro.food.navarrosfood.enums;
 
 import lombok.Getter;
