+++ Back end/src/main/java/com/navarro/food/navarrosfood/repositories/RepositoryFood.java
@@ -1,6 +1,6 @@
 package com.navarro.food.navarrosfood.repositories;
 
-import com.navarro.food.navarrosfood.model.Enums.Status;
+import com.navarro.food.navarrosfood.enums.Status;
 import com.navarro.food.navarrosfood.model.FoodEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
