+++ Back end/src/test/java/com/navarro/food/navarrosfood/Services/FoodServiceTest.java
@@ -1,10 +1,10 @@
 package com.navarro.food.navarrosfood.services;
 
 import com.navarro.food.navarrosfood.exception.FoodNotFound;
-import com.navarro.food.navarrosfood.model.DTOs.FoodRequest;
-import com.navarro.food.navarrosfood.model.DTOs.FoodResponse;
-import com.navarro.food.navarrosfood.model.DTOs.mapper.FoodMapper;
-import com.navarro.food.navarrosfood.model.Enums.Status;
+import com.navarro.food.navarrosfood.dtos.FoodRequest;
+import com.navarro.food.navarrosfood.dtos.FoodResponse;
+import com.navarro.food.navarrosfood.dtos.mapper.FoodMapper;
+import com.navarro.food.navarrosfood.enums.Status;
 import com.navarro.food.navarrosfood.model.FoodEntity;
 import com.navarro.food.navarrosfood.repositories.RepositoryFood;
 import com.navarro.food.navarrosfood.services.impl.ServiceFoodImpl;
